+++ src/components/Chat/Chatprofile.tsx
@@ -1,11 +1,15 @@
 import styled from "styled-components";
 import ChatproTitle from "./ChatproTitle";
+import Profile from "./Profile";
 
 const Chatprofile = () => {
 
     return (
         <ChatprofileWrapper>
             <ChatproTitle/>
+            <ProfileWrapper>
+                <Profile/>
+            </ProfileWrapper>
         </ChatprofileWrapper>
     );
 }
@@ -14,4 +18,8 @@ export default Chatprofile;
 
 const ChatprofileWrapper = styled.div`
     text-align: center;
+`;
+
+const ProfileWrapper = styled.div`
+    background-color:white;
 `;
